Refactor: fishable waters filters, doesn't filter with full text search
Instead, search-container.vue emits an object that is then parsed by
the map/index.vue page. The parsing of the object into a URL
querystring is handled with the URLSearchParams methods. When
passed an empty object it generates an empty string.

Waiting on water type classifications to include that in the
filter objecs/querstring.